refactor: move `manifest_and_lockfile` to shared `search` utility module
