+++ cli/flox/src/utils/mod.rs
@@ -10,5 +10,6 @@ pub mod init;
 pub mod logger;
 pub mod metrics;
 pub mod openers;
+pub mod search;
 
 pub static TERMINAL_STDERR: Lazy<Mutex<Stderr>> = Lazy::new(|| Mutex::new(std::io::stderr()));
